+++ server/runtime_go_nakama.go
@@ -2303,6 +2303,36 @@ func (n *RuntimeGoNakamaModule) LeaderboardRecordDelete(ctx context.Context, id,
 }
 
 // @group leaderboards
+// @summary Fetch the list of leaderboard records around the owner.
+// @param ctx(type=context.Context) The context object represents information about the server and requester.
+// @param id(type=string) The ID of the leaderboard to list records for.
+// @param ownerId(type=string) The owner ID around which to show records.
+// @param limit(type=int) Return only the required number of leaderboard records denoted by this limit value. Between 1-100.
+// @param expiry(type=int64) Time since epoch in seconds. Must be greater than 0.
+// @return leaderboardRecordsHaystack(*api.Leaderboard) A list of leaderboard records.
+// @return error(error) An optional error value if an error occurred.
+func (n *RuntimeGoNakamaModule) LeaderboardRecordsHaystack(ctx context.Context, id, ownerID string, limit int, expiry int64) ([]*api.LeaderboardRecord, error) {
+	if id == "" {
+		return nil, errors.New("expects a leaderboard ID string")
+	}
+
+	owner, err := uuid.FromString(ownerID)
+	if err != nil {
+		return nil, errors.New("expects owner ID to be a valid identifier")
+	}
+
+	if limit < 1 || limit > 100 {
+		return nil, errors.New("limit must be 1-100")
+	}
+
+	if expiry < 0 {
+		return nil, errors.New("expiry should be time since epoch in seconds and has to be a positive integer")
+	}
+
+	return LeaderboardRecordsHaystack(ctx, n.logger, n.db, n.leaderboardCache, n.leaderboardRankCache, id, owner, limit, expiry)
+}
+
+// @group leaderboards
 // @summary Fetch one or more leaderboards by ID.
 // @param ids(type=[]string) The table array of leaderboard ids.
 // @return leaderboardsGet(*api.Leaderboard) The leaderboard records according to ID.